requirements.txt
pandas
python-telegram-bot>=20.0
tabulate
